requirements.txt
fastapi
uvicorn
python-pptx
python-docx
Pillow
python-multipart
google-cloud-aiplatform
google-auth
